pyproject.toml
[tool.poetry]
name = "skypro-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
flake8 = "^7.1.1"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
markdown = "^3.6"
django-filter = "^24.3"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.42"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.1"
drf-yasg = "^1.21.7"
setuptools = "^73.0.1"
distribute = "^0.7.3"
django-cors-headers = "^4.4.0"
stripe = "^10.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
